Add plugin install command with shared update code
