+++ src/cli/Cargo.toml
@@ -28,7 +28,7 @@ rustc_version = "0.2.3"
 tracing = "0.1.19"
 tracing-futures = "0.2.4"
 url = "2.1.1"
-semver = "0.10.0"
+semver = "0.11.0"
 bytes = "0.5.3"
 structopt = { version = "0.3.16", default-features = false }
 toml = "0.5.5"
