+++ src/cli/src/lib.rs
@@ -14,7 +14,7 @@ mod tls;
 pub mod cluster;
 mod group;
 mod custom;
-mod update;
+mod install;
 mod partition;
 
 #[cfg(any(feature = "cluster_components", feature = "cluster_components_rustls"))]
